попытка разделить на файлы
